memisahkan script stack ke file lain
