+++ src/api/RancherProjectsApi.php
@@ -26,72 +26,4 @@ class RancherProjectsApi extends Controller
     return response()->json(Rancher::host()->all());
   }
 
-  public function listStackAll()
-  {
-    $stacks = Rancher::stack()->all();
-
-        foreach ($stacks as $stack)
-        {
-            unset($stack->dockerCompose);
-            unset($stack->rancherCompose);
-            unset($stack->healthState);
-            unset($stack->environment);
-            unset($stack->startOnCreate);
-            unset($stack->system);
-        
-        }
-        return response()->json($stacks);
-  }
-
-  public function listStackDB()
-  {
-    $listStackDB = RancherProjects::all();
-
-    $response["statusCode"] = 200;
-    $response["data"] = $listStackDB;
-
-    return response()->json($response);
-  }
-
-  public function cekStackDB(Request $request)
-  {
-    $id_stack     = $request->input('id_stack');
-
-    $listStackDB  = RancherProjects::where('rancher_stack_id', $id_stack)->first();
-
-    $response["statusCode"] = 200;
-    $response["data"] = $listStackDB;
-
-    return response()->json($response);
-  }
-
-  public function addStacktoDB(Request $request)
-  {
-    $id_stack = $request->input('stack_id');
-    $remark   = $request->input('remark');
-
-    $rancherProject = new RancherProjects();
-
-    $addStackDB = $rancherProject->simpan($id_stack, $remark);
-
-    $response["statusCode"] = 200;
-    $response["data"] = $addStackDB;
-
-    return response()->json($response);
-  }
-
-  public function deleteStackinDB(Request $request)
-  {
-    $id_stack = $request->input('stack_id');
-
-    $rancherProject = RancherProjects::where("rancher_stack_id", $id_stack);
-    $rancherProject->delete();
-
-    $response["statusCode"] = 200;
-    $response["data"] = $rancherProject;
-    return response()->json($response);
-
-  }
-
-
 }
