+++ src/components/RancherProjectsComponent.vue
@@ -88,7 +88,7 @@ export default {
   name: "RancherProjectsComponent",
   mounted() {
     this.instance = axios.create({
-      baseURL: '/tiketux/rancherprojects/api/'
+      baseURL: '/tiketux/rancher/stack/api/'
     });
     this.list();
   },
