+++ src/routes.php
@@ -1,8 +1,8 @@
 <?php
 
 Route::get('/tiketux/rancherprojects/api/list', 'Tiketux\RancherProjects\Api\RancherProjectsApi@listAll');
-Route::get('/tiketux/rancherprojects/api/liststack', 'Tiketux\RancherProjects\Api\RancherProjectsApi@listStackAll');
-Route::get('/tiketux/rancherprojects/api/liststackdb', 'Tiketux\RancherProjects\Api\RancherProjectsApi@listStackDB');
-Route::post('/tiketux/rancherprojects/api/addstackdb', 'Tiketux\RancherProjects\Api\RancherProjectsApi@addStacktoDB');
-Route::post('/tiketux/rancherprojects/api/cekstackdb', 'Tiketux\RancherProjects\Api\RancherProjectsApi@cekStackDB');
-Route::post('/tiketux/rancherprojects/api/deletestackdb', 'Tiketux\RancherProjects\Api\RancherProjectsApi@deleteStackinDB');
+Route::get('/tiketux/rancher/stack/api/liststack', 'Tiketux\RancherProjects\Api\RancherStackApi@listStackAll');
+Route::get('/tiketux/rancher/stack/api/liststackdb', 'Tiketux\RancherProjects\Api\RancherStackApi@listStackDB');
+Route::post('/tiketux/rancher/stack/api/addstackdb', 'Tiketux\RancherProjects\Api\RancherStackApi@addStacktoDB');
+Route::post('/tiketux/rancher/stack/api/cekstackdb', 'Tiketux\RancherProjects\Api\RancherStackApi@cekStackDB');
+Route::post('/tiketux/rancher/stack/api/deletestackdb', 'Tiketux\RancherProjects\Api\RancherStackApi@deleteStackinDB');
